Don't add to a project board or trigger any workflows on creation (#54)
We used to track upgrades by adding them to the a GH project:
https://github.com/orgs/pulumi/projects/28#card-91004424. We don't do
this anymore, so we
don't need to keep the code to add it.

We used to run an automated workflow: "update-upstream-provider.yml", on
issue
creation. This workflow no longer exists. We run the modern equivalent
by triggering on issue creation.

We can simplify this project by removing both of these pieces of
automation.